Login endpoint for OpenIdConnect flow (#59)
This endpoint is intended for browsers only.
It will trigger DefaultChallengeScheme, the authentication process using OpenIdConnect flow if the user is not authenticated.
It will then set a cookie upon successful authentication.
From this point on, the endpoints that use Cookies authentication scheme will work.
The endpoints that allow both Cookies and Bearer authentication schemes only work if the user has a valid cookie or a valid Bearer token.
They will not trigger the authentication process.